# Improve performance of algorithm `HyperND`.