@GiggleLiu 2025-07-09
I agree there should be something like the `cost` like you proposed. Actually, we already have a similar parameters in `TreeSA`, called `tc_weight` and `rw_weight`.

After another PR #75 that refactors TreeSA being merged. I will try to propose some consistent ways to specify the cost.